requirements.txt
asgiref==3.11.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
factory_boy==3.3.3
Faker==38.2.0
gunicorn==21.2.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
pytest-mock==3.15.1
pytz==2025.2
sqlparse==0.5.4
tzdata==2025.2
